refactor: move commands and results to `dto` package for improved project structure
