+++ src/main/java/com/yeoljeong/tripmate/product/application/dto/command/CreateProductCommand.java
@@ -1,4 +1,4 @@
-package com.yeoljeong.tripmate.product.application.command;
+package com.yeoljeong.tripmate.product.application.dto.command;
 
 import com.yeoljeong.tripmate.product.domain.entity.Product;
 import com.yeoljeong.tripmate.product.domain.entity.ProductAddress;
