+++ src/main/java/com/yeoljeong/tripmate/product/presentation/dto/response/ProductResponse.java
@@ -1,6 +1,6 @@
 package com.yeoljeong.tripmate.product.presentation.dto.response;
 
-import com.yeoljeong.tripmate.product.application.result.ProductResult;
+import com.yeoljeong.tripmate.product.application.dto.result.ProductResult;
 import com.yeoljeong.tripmate.product.domain.entity.ProductAddress;
 import java.math.BigDecimal;
 import java.util.UUID;
